uboot-lantiq: add support for ZTE H201L
Support for this board was dropped by me in assumption this board isn't
supported by OpenWrt at all. Turns out, this board is supported by the
lantiq target and a custom bootloader is required for the OpenWrt image.

Add the board again and apply all changes applied for the other boards
in the meantime.

The RAM settings extracted from the ZTE bootloader are obviously tuned
ones, which do not work with the UART/RAM bootloader. Instead not
optimized RAM setting are added and only used for the UART/RAM
bootloader.

Signed-off-by: Mathias Kresin <dev@kresin.me>